Improve example systemd service file
Remove set-environment.service as it is something specific to a project where we use this. This example relies on /etc/manage-ssl-cert instead, which could be generated by another service but that is left as an exercise for the reader.

Added RemainAfterExit=no to make it explicit that it should not be set to yes, as that would interfere with the timer.

Changed RequiredBy=nginx.service to WantedBy=nginx.service because RequiredBy tells systemd to stop nginx when this service is stopped, which is not desired.